moved int conversion from data reading in data.py to the defiinition of beginning times in inference.py

+++ soundbay/data.py
@@ -360,7 +360,7 @@ class InferenceDataset(Dataset):
             audio_dict contains references to audio paths given name from metadata
         """
         audio_len = sf.info(self.file_path).duration
-        self._start_times = np.arange(0, audio_len//self.seq_length * self.seq_length, self.seq_length).astype(int)
+        self._start_times = np.arange(0, audio_len//self.seq_length * self.seq_length, self.seq_length)
 
     def _get_audio(self, begin_time):
         """
@@ -393,7 +393,7 @@ class InferenceDataset(Dataset):
         audio, label - torch tensor (1-d if no spectrogram is applied/ 2-d if applied a spectrogram
         , int (if mode="train" only)
         '''
-        begin_time = self._start_times[idx] * self.data_sample_rate
+        begin_time = int(self._start_times[idx] * self.data_sample_rate)
         audio = self._get_audio(begin_time)
         audio = self.sampler(audio)
         audio = self.preprocessor(audio)
